# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取我的信息",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/update": {
      "post": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新我的资料（POST 语义：执行一次更新动作）",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/password": {
      "post": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改密码（旧密码校验 + pv 自增）",
        "tags": [
          "users"
        ]
      }
    },
...
    "/article/list": {
      "get": {
        "description": "获取已发布的文章列表，支持分页和筛选",
        "operationId": "PublicArticleController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数，默认 20，最大 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字搜索（标题、摘要）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "分类 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回文章列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticleListVO"
                }
              }
            }
          }
        },
        "summary": "获取文章列表",
        "tags": [
          "文章 - 用户端"
        ]
      }
    },
    "/article/featured": {
      "get": {
        "description": "获取推荐的文章列表",
        "operationId": "PublicArticleController_findFeatured",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "数量限制，默认10",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回推荐文章列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicArticleListItemVO"
                  }
                }
              }
            }
          }
        },
        "summary": "获取推荐文章",
        "tags": [
          "文章 - 用户端"
        ]
      }
    },
    "/article/top": {
      "get": {
        "description": "获取置顶的文章列表",
        "operationId": "PublicArticleController_findTop",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "数量限制，默认5",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回置顶文章列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicArticleListItemVO"
                  }
                }
              }
            }
          }
        },
        "summary": "获取置顶文章",
        "tags": [
          "文章 - 用户端"
        ]
      }
    },
    "/article/detail/{uid}": {
      "get": {
        "description": "根据 UID 获取文章详情，并增加阅读量",
        "operationId": "PublicArticleController_findOne",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "文章 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回文章详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticleDetailVO"
                }
              }
            }
          }
        },
        "summary": "获取文章详情",
        "tags": [
          "文章 - 用户端"
        ]
      }
    },
    "/article/like/{uid}": {
      "post": {
        "description": "为文章点赞",
        "operationId": "PublicArticleController_like",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "文章 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回当前点赞数"
          }
        },
        "summary": "点赞文章",
        "tags": [
          "文章 - 用户端"
        ]
      }
    },
    "/article/category/{categoryId}": {
      "get": {
        "description": "获取指定分类下的文章列表",
        "operationId": "PublicArticleController_findByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数，默认 20，最大 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字搜索（标题、摘要）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回文章列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticleListVO"
                }
              }
            }
          }
        },
        "summary": "按分类获取文章",
        "tags": [
          "文章 - 用户端"
        ]
      }
    },
...
    "/admin/users/create": {
      "post": {
        "operationId": "AdminUsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建管理员（POST）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminUsersController_list",
        "parameters": [
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "按 username/email/nickname 模糊搜索",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "active|inactive|banned",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码（默认1）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数（默认20）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "管理员列表（GET，支持分页/搜索）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/detail": {
      "post": {
        "operationId": "AdminUsersController_detail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "管理员详情（POST）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/update": {
      "post": {
        "operationId": "AdminUsersController_update",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新管理员资料（POST，目标管理员）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/delete": {
      "post": {
        "operationId": "AdminUsersController_remove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除管理员（POST）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/batch-delete": {
      "post": {
        "operationId": "AdminUsersController_batchRemove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量删除管理员（POST）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/assign-roles": {
      "post": {
        "operationId": "AdminUsersController_assignRoles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "为管理员分配角色（POST，按 roleCodes 全量覆盖）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/me": {
      "get": {
        "operationId": "AdminUsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "我的信息（GET）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/me/update": {
      "post": {
        "operationId": "AdminUsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新我的资料（POST）",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/admin/users/me/password": {
      "post": {
        "operationId": "AdminUsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改我的密码（POST）",
        "tags": [
          "admin-users"
        ]
      }
    },
...
  },
  "info": {
    "title": "jmni-server",
    "description": "JMNI Server API 文档 - 供外部团队参考",
    "version": "0.0.1",
    "contact": {
      "name": "JMNI Team",
      "url": "https://jmni.cn",
      "email": "support@jmni.cn"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://api.jmni.cn",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:2233",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Access Token"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "昵称",
            "maxLength": 50
          },
          "avatar_url": {
            "type": "string",
            "description": "头像 URL"
          },
          "gender": {
            "type": "string",
            "description": "性别",
            "enum": [
              "unknown",
              "male",
              "female",
              "other"
            ]
          },
          "birthday": {
            "type": "string",
            "description": "生日（YYYY-MM-DD）",
            "example": "1990-06-01"
          },
          "country": {
            "type": "string",
            "description": "国家/地区码（ISO 3166-1 alpha-2）",
            "example": "CN"
          },
          "locale": {
            "type": "string",
            "description": "语言（如 zh-CN）",
            "example": "zh-CN",
            "maxLength": 10
          },
          "time_zone": {
            "type": "string",
            "description": "时区（IANA）",
            "example": "Asia/Shanghai",
            "maxLength": 40
          },
          "marketing_consent": {
            "type": "boolean",
            "description": "营销/通知同意",
            "example": false
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "current": {
            "type": "string",
            "description": "当前密码",
            "example": "Old@123456"
          },
          "new": {
            "type": "string",
            "description": "新密码",
            "example": "New@123456"
          }
        },
        "required": [
          "current",
          "new"
        ]
      },
...
      "PublicArticleListItemVO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "description": "文章 UID",
            "example": "123456789012"
          },
          "title": {
            "type": "string",
            "description": "标题",
            "example": "如何使用 NestJS"
          },
          "subTitle": {
            "type": "string",
            "description": "子标题"
          },
          "summary": {
            "type": "string",
            "description": "摘要",
            "example": "本文介绍..."
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图 URL"
          },
          "categoryId": {
            "type": "number",
            "description": "分类 ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称",
            "example": "技术文章"
          },
          "tags": {
            "type": "array",
            "description": "标签",
            "items": {
              "type": "string"
            }
          },
          "isTop": {
            "type": "boolean",
            "description": "是否置顶",
            "example": false
          },
          "isFeatured": {
            "type": "boolean",
            "description": "是否推荐",
            "example": false
          },
          "viewCount": {
            "type": "number",
            "description": "阅读量",
            "example": 100
          },
          "likeCount": {
            "type": "number",
            "description": "点赞数",
            "example": 10
          },
          "publishedAt": {
            "type": "string",
            "description": "发布时间（ISO 8601）"
          }
        },
        "required": [
          "uid",
          "title",
          "summary",
          "categoryName",
          "isTop",
          "isFeatured",
          "viewCount",
          "likeCount"
        ]
      },
      "PublicArticleListVO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "总记录数",
            "example": 100
          },
          "items": {
            "description": "文章列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicArticleListItemVO"
            }
          },
          "page": {
            "type": "number",
            "description": "当前页码",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "description": "总页数",
            "example": 5
          }
        },
        "required": [
          "total",
          "items",
          "page",
          "pageSize",
          "totalPages"
        ]
      },
      "PublicArticleDetailVO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "description": "文章 UID",
            "example": "123456789012"
          },
          "title": {
            "type": "string",
            "description": "标题",
            "example": "如何使用 NestJS"
          },
          "subTitle": {
            "type": "string",
            "description": "子标题"
          },
          "summary": {
            "type": "string",
            "description": "摘要",
            "example": "本文介绍..."
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图 URL"
          },
          "categoryId": {
            "type": "number",
            "description": "分类 ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称",
            "example": "技术文章"
          },
          "tags": {
            "type": "array",
            "description": "标签",
            "items": {
              "type": "string"
            }
          },
          "isTop": {
            "type": "boolean",
            "description": "是否置顶",
            "example": false
          },
          "isFeatured": {
            "type": "boolean",
            "description": "是否推荐",
            "example": false
          },
          "viewCount": {
            "type": "number",
            "description": "阅读量",
            "example": 100
          },
          "likeCount": {
            "type": "number",
            "description": "点赞数",
            "example": 10
          },
          "publishedAt": {
            "type": "string",
            "description": "发布时间（ISO 8601）"
          },
          "content": {
            "type": "string",
            "description": "文章内容"
          },
          "contentFormat": {
            "type": "string",
            "description": "内容格式",
            "enum": [
              "markdown",
              "html",
              "text"
            ],
            "example": "markdown"
          },
          "seoTitle": {
            "type": "string",
            "description": "SEO 标题"
          },
          "seoKeywords": {
            "type": "string",
            "description": "SEO 关键词"
          },
          "seoDescription": {
            "type": "string",
            "description": "SEO 描述"
          }
        },
        "required": [
          "uid",
          "title",
          "summary",
          "categoryName",
          "isTop",
          "isFeatured",
          "viewCount",
          "likeCount",
          "content",
          "contentFormat",
          "seoTitle",
          "seoKeywords",
          "seoDescription"
        ]
      },
...
      "AdminCreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "john_doe",
            "minLength": 3,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "Password123!",
            "minLength": 6,
            "maxLength": 128
          },
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "john@example.com"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "maxLength": 50
          },
          "emailcode": {
            "type": "string",
            "description": "注册用邮箱验证码（有 email 时必填）",
            "minLength": 4,
            "maxLength": 8
          },
          "roleCodes": {
            "description": "角色代码列表（可选，按 code 绑定）",
            "example": [
              "super_admin",
              "ops"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "password"
        ]
      },
      "IdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "主键ID",
            "example": 1
          }
        },
        "required": [
          "id"
        ]
      },
      "IdsDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "主键ID列表",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "AssignRolesDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "管理员ID",
            "example": 1
          },
          "roleCodes": {
            "description": "角色代码数组（全量覆盖）",
            "example": [
              "super_admin",
              "editor"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "roleCodes"
        ]
      },
      "AdminUpdateUserDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "status": {
            "type": "string",
            "description": "状态",
            "enum": [
              "active",
              "inactive",
              "banned"
            ]
          }
        }
      },
...
    }
  }
}